resume {resume}	R Documentation

Template to build Resume.

Description

First create a new R Markdown document from template Resume then modify it and knit PDF output.

Usage


resume(..., keep_tex = TRUE)

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

Value

R Markdown output format to pass to render

Examples



## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "resume", package = "resume")

## End(Not run)


[Package resume version 0.2 Index]